backend/requirements.txt
pydantic
fastapi
numpy
opencv-python
Pillow
ultralytics
paddlepaddle
paddleocr